+++ tuner_hal2/aidl_service/src/service_entry.rs
@@ -661,7 +661,9 @@ pub fn run_service() {
     {
         std::process::exit(1);
     }
-    if crate::key_provisioning_bridge_server::start_key_provisioning_bridge_server(context.clone()).is_err() {
+    if crate::key_provisioning_bridge_server::start_key_provisioning_server(context.clone())
+        .is_err()
+    {
         std::process::exit(1);
     }
     let tuner = match TunerAidlService::from_context(context) {